Переезд сотрудников в отдельный контроллер. Добавление и редактирование пунктов меню
