+++ WebStore/Views/Employees/Details.cshtml
@@ -1,0 +1,23 @@
+﻿@model WebStore.Models.Employee
+
+    <div class="container">
+        <h3>Сотрудник</h3>
+        <dl class="dl-horizontal">
+            <dt>Id</dt>
+            <dd>@Model.Id</dd>
+
+            <dt>Фамилия</dt>
+            <dd>@Model.Surname</dd>
+
+            <dt>Имя</dt>
+            <dd>@Model.Name</dd>
+
+            <dt>Отчество</dt>
+            <dd>@Model.Patronymic</dd>
+
+            <dt>Возраст</dt>
+            <dd>@Model.Age</dd>
+        </dl>
+        <hr />
+        <a asp-controller="Employees" asp-action="Index">К списку сотрудников</a>
+    </div>
