+++ WebStore/Views/Shared/Partial/_MainMenu.cshtml
@@ -1,22 +1,22 @@
 ﻿<ul class="nav navbar-nav collapse navbar-collapse">
-    <li><a href="index.html" class="active">Home</a></li>
+    <li><a asp-controller="Home" asp-action="Index" class="active">Главная</a></li>
     <li class="dropdown">
-        <a href="#">Shop<i class="fa fa-angle-down"></i></a>
+        <a href="#">Магазин<i class="fa fa-angle-down"></i></a>
         <ul role="menu" class="sub-menu">
-            <li><a href="shop.html">Products</a></li>
-            <li><a href="product-details.html">Product Details</a></li>
-            <li><a href="checkout.html">Checkout</a></li>
-            <li><a href="cart.html">Cart</a></li>
-            <li><a href="login.html">Login</a></li>
+            <li><a href="shop.html">Каталог товаров</a></li>
+            <li><a href="product-details.html">Информация о товаре</a></li>
+            <li><a href="checkout.html">Оформление заказа</a></li>
+            <li><a href="cart.html">Корзина</a></li>
         </ul>
     </li>
     <li class="dropdown">
-        <a href="#">Blog<i class="fa fa-angle-down"></i></a>
+        <a href="#">Блоги<i class="fa fa-angle-down"></i></a>
         <ul role="menu" class="sub-menu">
-            <li><a href="blog.html">Blog List</a></li>
-            <li><a href="blog-single.html">Blog Single</a></li>
+            <li><a href="blog.html">Список блогов</a></li>
+            <li><a href="blog-single.html">Блог магазина</a></li>
         </ul>
     </li>
-    <li><a href="404.html">404</a></li>
-    <li><a href="contact-us.html">Contact</a></li>
+    <li><a href="404.html">Ошибка 404</a></li>
+    <li><a href="contact-us.html">Контакты</a></li>
+    <li><a asp-controller="Employees" asp-action="Index">Сотрудники</a></li>
 </ul>
